pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "predx"
version = "0.1.0"
description = "Python library for researching and trading on Kalshi and Polymarket"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "cryptography>=42.0",
    "websockets>=12.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
poly-trading = ["py-clob-client>=0.17"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.hatch.build.targets.wheel]
packages = ["predx"]
